@@ -61,69 +65,72 @@
     "modern-errors": "6.0.0",
     "modern-errors-http": "4.0.0",
     "modern-errors-serialize": "5.0.0",
-    "next": "13.4.4",
+    "next": "13.4.19",
     "omit-deep-lodash": "1.1.7",
-    "pino": "8.14.1",
-    "pino-pretty": "10.0.0",
+    "pino": "8.15.0",
+    "pino-pretty": "10.2.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-hook-form": "7.44.2",
+    "react-hook-form": "7.45.4",
+    "semver": "7.5.4",
+    "tsx": "3.12.7",
     "url-join": "5.0.0",
     "urql": "3.0.4",
-    "uuidv7": "0.4.3",
-    "vite": "4.3.9",
-    "vitest": "0.31.1",
-    "webpack": "5.84.1",
+    "uuidv7": "0.6.2",
+    "vite": "4.4.9",
+    "vitest": "0.34.2",
+    "webpack": "5.88.2",
     "yup": "1.2.0",
-    "zod": "3.21.4",
-    "zustand": "4.3.8"
+    "zod": "3.22.2",
+    "zustand": "4.4.1"
   },
   "devDependencies": {
     "@graphql-codegen/add": "5.0.0",
-    "@graphql-codegen/cli": "4.0.0",
+    "@graphql-codegen/cli": "5.0.0",
     "@graphql-codegen/introspection": "4.0.0",
-    "@graphql-codegen/typed-document-node": "5.0.0",
-    "@graphql-codegen/typescript": "4.0.0",
-    "@graphql-codegen/typescript-operations": "4.0.0",
+    "@graphql-codegen/typed-document-node": "5.0.1",
+    "@graphql-codegen/typescript": "4.0.1",
+    "@graphql-codegen/typescript-operations": "4.0.1",
     "@graphql-codegen/typescript-urql": "3.7.3",
     "@graphql-codegen/urql-introspection": "2.2.1",
     "@graphql-typed-document-node/core": "3.2.0",
-    "@next/env": "13.4.4",
-    "@pollyjs/adapter-fetch": "6.0.5",
-    "@pollyjs/adapter-node-http": "6.0.5",
-    "@pollyjs/core": "6.0.5",
-    "@pollyjs/persister-fs": "6.0.5",
+    "@pollyjs/adapter-fetch": "6.0.6",
+    "@pollyjs/adapter-node-http": "6.0.6",
+    "@pollyjs/core": "6.0.6",
+    "@pollyjs/persister-fs": "6.0.6",
     "@saleor/eslint-plugin-saleor-app": "0.1.2",
-    "@testing-library/jest-dom": "5.16.5",
+    "@saleor/json-schema-compiler": "0.1.2",
+    "@testing-library/jest-dom": "6.1.1",
     "@testing-library/react": "14.0.0",
     "@testing-library/react-hooks": "8.0.1",
     "@types/bluebird": "3.5.38",
-    "@types/lodash-es": "4.17.7",
-    "@types/node": "20.2.5",
+    "@types/lodash-es": "4.17.8",
+    "@types/node": "20.5.4",
     "@types/omit-deep-lodash": "1.1.1",
-    "@types/react": "18.2.7",
-    "@types/react-dom": "18.2.4",
+    "@types/react": "18.2.21",
+    "@types/react-dom": "18.2.7",
+    "@types/semver": "7.5.0",
     "@types/setup-polly-jest": "0.5.2",
-    "@types/testing-library__jest-dom": "5.14.6",
-    "@typescript-eslint/eslint-plugin": "5.59.8",
-    "@typescript-eslint/parser": "5.59.8",
-    "@vanilla-extract/vite-plugin": "3.8.2",
-    "dependency-cruiser": "13.0.2",
-    "eslint": "8.41.0",
-    "eslint-config-next": "13.4.4",
-    "eslint-config-prettier": "8.8.0",
-    "eslint-plugin-import": "2.27.5",
+    "@typescript-eslint/eslint-plugin": "6.4.1",
+    "@typescript-eslint/parser": "6.4.1",
+    "@vanilla-extract/vite-plugin": "3.9.0",
+    "@vitest/coverage-v8": "0.34.2",
+    "dependency-cruiser": "13.1.4",
+    "eslint": "8.47.0",
+    "eslint-config-next": "13.4.19",
+    "eslint-config-prettier": "9.0.0",
+    "eslint-plugin-import": "2.28.1",
     "eslint-plugin-require-form-method": "1.0.2",
-    "eslint-plugin-testing-library": "5.11.0",
-    "eslint-plugin-vitest": "0.2.5",
+    "eslint-plugin-testing-library": "6.0.0",
+    "eslint-plugin-vitest": "0.2.8",
     "husky": "8.0.3",
-    "json-schema-to-typescript": "13.0.1",
-    "lint-staged": "13.2.2",
+    "json-schema-to-typescript": "13.0.2",
+    "lint-staged": "14.0.1",
     "next-test-api-route-handler": "3.1.8",
-    "prettier": "2.8.8",
+    "prettier": "3.0.2",
     "setup-polly-jest": "0.11.0",
     "ts-node": "10.9.1",
-    "typescript": "5.0.4",
+    "typescript": "5.1.6",
     "vite-tsconfig-paths": "4.2.0"
   },
   "engines": {
