@@ -1,8 +1,9 @@
 import "next";
 import { loadEnvConfig } from "@next/env";
 import { expect, afterEach } from "vitest";
-import matchers from "@testing-library/jest-dom/matchers";
+import * as matchers from "@testing-library/jest-dom/matchers";
 import { cleanup } from "@testing-library/react";
+import "@testing-library/jest-dom/vitest";
 
 loadEnvConfig(".");
 
